Add detect worker connected

+++ packages/think-cluster/lib/master.js
@@ -78,6 +78,7 @@ class Master {
     worker.kill('SIGINT'); //windows don't support SIGQUIT
     if(reload) worker.hasGracefulReload = true;
     setTimeout(function () {
+      if(!worker.isConnected()) return;
       worker.process.kill('SIGINT');
     }, 100);
   }
